move intermediate ops and sink to separate files
