+++ src/daft-local-execution/src/pipeline.rs
@@ -1,4 +1,4 @@
-use std::{borrow::BorrowMut, pin::Pin, sync::Arc};
+use std::{pin::Pin, sync::Arc};
 
 use common_error::{DaftError, DaftResult};
 use daft_micropartition::MicroPartition;
@@ -10,10 +10,10 @@ use tokio_stream::wrappers::ReceiverStream;
 
 use crate::{
     create_channel,
-    intermediate_ops::intermediate_op::IntermediateOperator,
-    sinks::sink::{Sink, SinkResultType},
+    intermediate_ops::intermediate_op::{run_intermediate_operators, IntermediateOperator},
+    sinks::sink::{launch_sink, Sink},
     sources::source::Source,
-    Receiver, Sender,
+    Sender,
 };
 
 lazy_static! {
@@ -46,101 +46,6 @@ impl Pipeline {
         self
     }
 
-    async fn run_intermediate_operators(
-        mut receiver: Receiver,
-        intermediate_operators: Vec<Box<dyn IntermediateOperator>>,
-        send_to_sink: Sender,
-    ) -> DaftResult<()> {
-        log::debug!(
-            "Running intermediate operators: {}",
-            intermediate_operators
-                .iter()
-                .fold(String::new(), |acc, op| { acc + &op.name() + " -> " })
-        );
-
-        while let Some(morsel) = receiver.recv().await {
-            let mut data = morsel?;
-            for op in intermediate_operators.iter() {
-                data = op.execute(&data)?;
-            }
-            let _ = send_to_sink.send(Ok(data)).await;
-        }
-
-        log::debug!("Intermediate operators finished");
-        Ok(())
-    }
-
-    async fn launch_sink(
-        mut receivers: Vec<Receiver>,
-        mut sink: Option<Box<dyn Sink>>,
-        send_to_next_source: Sender,
-    ) -> DaftResult<()> {
-        let in_order = match sink {
-            Some(ref sink) => sink.in_order(),
-            None => false,
-        };
-        log::debug!("Launching sink with in_order: {}", in_order);
-
-        if in_order {
-            let mut finished_receiver_idxs = std::collections::HashSet::new();
-            let mut curr_idx = 0;
-            while finished_receiver_idxs.len() != receivers.len() {
-                if finished_receiver_idxs.contains(&curr_idx) {
-                    curr_idx = (curr_idx + 1) % receivers.len();
-                    continue;
-                }
-                let receiver = receivers.get_mut(curr_idx).expect("Receiver not found");
-                if let Some(val) = receiver.recv().await {
-                    if let Some(sink) = sink.borrow_mut() {
-                        let sink_result = sink.sink(&val?)?;
-                        match sink_result {
-                            SinkResultType::Finished => {
-                                break;
-                            }
-                            SinkResultType::NeedMoreInput => {}
-                        }
-                    } else {
-                        log::debug!("No sink, sending value to next source");
-                        let _ = send_to_next_source.send(val).await;
-                    }
-                } else {
-                    finished_receiver_idxs.insert(curr_idx);
-                }
-                curr_idx = (curr_idx + 1) % receivers.len();
-            }
-        } else {
-            let mut unordered_receivers = FuturesUnordered::new();
-            for receiver in receivers.iter_mut() {
-                unordered_receivers.push(receiver.recv());
-            }
-            while let Some(val) = unordered_receivers.next().await {
-                if let Some(val) = val {
-                    if let Some(sink) = sink.borrow_mut() {
-                        let sink_result = sink.sink(&val?)?;
-                        match sink_result {
-                            SinkResultType::Finished => {
-                                break;
-                            }
-                            SinkResultType::NeedMoreInput => {}
-                        }
-                    } else {
-                        log::debug!("No sink, sending value to next source");
-                        let _ = send_to_next_source.send(val).await;
-                    }
-                }
-            }
-        }
-
-        if let Some(sink) = sink.borrow_mut() {
-            let final_values = sink.finalize()?;
-            for value in final_values {
-                log::debug!("Sending finalized value to next source");
-                let _ = send_to_next_source.send(Ok(value)).await;
-            }
-        }
-        Ok(())
-    }
-
     pub async fn run(
         send_to_next_source: Sender,
         source: Box<dyn Source>,
@@ -163,7 +68,7 @@ impl Pipeline {
             } else {
                 let (tx1, rx1) = create_channel();
                 let (tx2, rx2) = create_channel();
-                let handle = tokio::spawn(Self::run_intermediate_operators(
+                let handle = tokio::spawn(run_intermediate_operators(
                     rx1,
                     intermediate_operators.clone(),
                     tx2,
@@ -176,8 +81,7 @@ impl Pipeline {
             curr_idx = (curr_idx + 1) % intermediate_senders_and_handles.len();
         }
 
-        let sink_handle =
-            tokio::spawn(Self::launch_sink(sink_receivers, sink, send_to_next_source));
+        let sink_handle = tokio::spawn(launch_sink(sink_receivers, sink, send_to_next_source));
 
         let mut handles_unordered = FuturesUnordered::new();
         for (sender, handle) in intermediate_senders_and_handles {
