+++ src/daft-local-execution/src/sinks/mod.rs
@@ -1,2 +1,3 @@
+pub mod aggregate;
 pub mod limit;
 pub mod sink;
